Inspection of overflow (see description)
There was an overflow with estimator 'smooth_kernel' especially when considering the Exponential DGP. This is not a bug. The standard error is very large, and it comes from the 'inverse of the density' terms that goes into the standard error. With that DGP, some points are very extreme, so the density estimator becomes very close to zero at the right tail, which becomes a problem when a \hat U is large and falls, e.g., between two extreme values.